refactor: フォルダー名を home に設定
